+++ src/com/gregswebserver/ld28/game/level/tile/WallLandmarkTile.java
@@ -1,9 +1,9 @@
 package com.gregswebserver.ld28.game.level.tile;
 
-import com.gregswebserver.ld28.util.vectors.Vector2i;
+import com.gregswebserver.ld28.util.vectors.Vector2d;
 
 public class WallLandmarkTile extends WallTile {
-    public WallLandmarkTile(Vector2i position) {
+    public WallLandmarkTile(Vector2d position) {
         super(position);
     }
 }
