+++ src/com/gregswebserver/ld28/input/KeyboardHandler.java
@@ -1,7 +1,7 @@
 package com.gregswebserver.ld28.input;
 
 import com.gregswebserver.ld28.game.UsesGame;
-import com.gregswebserver.ld28.util.vectors.Vector2i;
+import com.gregswebserver.ld28.util.vectors.Vector2d;
 
 import java.awt.event.KeyEvent;
 import java.awt.event.KeyListener;
@@ -27,12 +27,12 @@ public class KeyboardHandler extends UsesGame implements KeyListener {
         return keys.contains("" + key);
     }
 
-    public Vector2i getArrows() {
-        Vector2i sum = new Vector2i();
-        if (isPressed(KeyEvent.VK_UP) || isPressed(KeyEvent.VK_W)) sum.add(new Vector2i(0,1));
-        if (isPressed(KeyEvent.VK_RIGHT) || isPressed(KeyEvent.VK_D)) sum.add(new Vector2i(1,0));
-        if (isPressed(KeyEvent.VK_DOWN) || isPressed(KeyEvent.VK_S)) sum.add(new Vector2i(0,-1));
-        if (isPressed(KeyEvent.VK_LEFT) || isPressed(KeyEvent.VK_A)) sum.add(new Vector2i(-1,0));
+    public Vector2d getArrows() {
+        Vector2d sum = new Vector2d();
+        if (isPressed(KeyEvent.VK_UP) || isPressed(KeyEvent.VK_W)) sum.add(new Vector2d(0, 1));
+        if (isPressed(KeyEvent.VK_RIGHT) || isPressed(KeyEvent.VK_D)) sum.add(new Vector2d(1, 0));
+        if (isPressed(KeyEvent.VK_DOWN) || isPressed(KeyEvent.VK_S)) sum.add(new Vector2d(0, -1));
+        if (isPressed(KeyEvent.VK_LEFT) || isPressed(KeyEvent.VK_A)) sum.add(new Vector2d(-1, 0));
         return sum;
     }
 }
